# openapi.json
{
  "components": {
    "schemas": {
      "AcceptInvitationResponseDto": {
        "properties": {
          "eventId": {
            "description": "Event the user is now a planning member of.",
            "example": "6bb69700-72eb-40a0-b37e-e6c0b8b004a9",
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "eventId"
        ],
        "type": "object"
      },
...
      "InviteMemberDto": {
        "properties": {
          "email": {
            "example": "tunde@example.com",
            "maxLength": 255,
            "type": "string"
          },
          "role": {
            "enum": [
              "admin",
              "contributor",
              "observer"
            ],
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "type": "object"
      },
      "InviteMemberResponseDto": {
        "properties": {
          "acceptedAt": {
            "description": "Timestamp the recipient accepted; null while pending/declined/revoked/expired.",
            "example": null,
            "nullable": true,
            "type": "string"
          },
          "acceptUrl": {
            "description": "Full URL the recipient can use to accept. Same token is in the email; admins copy this to share manually.",
            "example": "https://app.epdema.com/invitations/3f9c1e7b8a2d4f5e6c0b9a1d2e3f4a5b",
            "type": "string"
          },
          "createdAt": {
            "example": "2026-06-02T08:12:00.000Z",
            "type": "string"
          },
          "email": {
            "example": "bola@example.com",
            "type": "string"
          },
          "emailError": {
            "description": "Error message if email delivery failed; absent on success.",
            "example": "Resend sandbox: recipient not on allowlist",
            "type": "string"
          },
          "emailSent": {
            "description": "Whether the invitation email was sent successfully (best-effort).",
            "example": true,
            "type": "boolean"
          },
          "eventId": {
            "example": "6bb69700-72eb-40a0-b37e-e6c0b8b004a9",
            "format": "uuid",
            "type": "string"
          },
          "expiresAt": {
            "example": "2026-06-09T08:12:00.000Z",
            "type": "string"
          },
          "id": {
            "example": "b2c3d4e5-f6a7-4b8c-9d0e-1f2a3b4c5d6e",
            "format": "uuid",
            "type": "string"
          },
          "inviterId": {
            "description": "User id of the admin who sent the invitation.",
            "example": "usr_1A2b3C4d5E6f7G8h9I0jKl",
            "type": "string"
          },
          "role": {
            "enum": [
              "admin",
              "contributor",
              "observer"
            ],
            "example": "contributor",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "accepted",
              "declined",
              "revoked",
              "expired"
            ],
            "example": "pending",
            "type": "string"
          }
        },
        "required": [
          "id",
          "eventId",
          "email",
          "role",
          "status",
          "expiresAt",
          "inviterId",
          "createdAt",
          "acceptUrl",
          "emailSent"
        ],
        "type": "object"
      },
...
      "PlanningInvitationDto": {
        "properties": {
          "acceptedAt": {
            "description": "Timestamp the recipient accepted; null while pending/declined/revoked/expired.",
            "example": null,
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "example": "2026-06-02T08:12:00.000Z",
            "type": "string"
          },
          "email": {
            "example": "bola@example.com",
            "type": "string"
          },
          "eventId": {
            "example": "6bb69700-72eb-40a0-b37e-e6c0b8b004a9",
            "format": "uuid",
            "type": "string"
          },
          "expiresAt": {
            "example": "2026-06-09T08:12:00.000Z",
            "type": "string"
          },
          "id": {
            "example": "b2c3d4e5-f6a7-4b8c-9d0e-1f2a3b4c5d6e",
            "format": "uuid",
            "type": "string"
          },
          "inviterId": {
            "description": "User id of the admin who sent the invitation.",
            "example": "usr_1A2b3C4d5E6f7G8h9I0jKl",
            "type": "string"
          },
          "role": {
            "enum": [
              "admin",
              "contributor",
              "observer"
            ],
            "example": "contributor",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "accepted",
              "declined",
              "revoked",
              "expired"
            ],
            "example": "pending",
            "type": "string"
          }
        },
        "required": [
          "id",
          "eventId",
          "email",
          "role",
          "status",
          "expiresAt",
          "inviterId",
          "createdAt"
        ],
        "type": "object"
      },
      "PlanningMemberDto": {
        "properties": {
          "eventId": {
            "example": "6bb69700-72eb-40a0-b37e-e6c0b8b004a9",
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "example": "a3f1d9c8-4b2e-4f6a-9c1d-7e8f0a2b3c4d",
            "format": "uuid",
            "type": "string"
          },
          "invitedById": {
            "description": "The member who invited this user. Null for the event creator (auto-added as first admin).",
            "example": "usr_1A2b3C4d5E6f7G8h9I0jKl",
            "nullable": true,
            "type": "string"
          },
          "joinedAt": {
            "example": "2026-05-26T08:12:00.000Z",
            "type": "string"
          },
          "role": {
            "enum": [
              "admin",
              "contributor",
              "observer"
            ],
            "example": "contributor",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/PlanningMemberUserDto"
          },
          "userId": {
            "example": "usr_2N9k8xQpL3vM5Rb1Tw7yZc",
            "type": "string"
          }
        },
        "required": [
          "id",
          "eventId",
          "userId",
          "role",
          "joinedAt",
          "user"
        ],
        "type": "object"
      },
      "PlanningMemberUserDto": {
        "properties": {
          "email": {
            "example": "tunde@example.com",
            "type": "string"
          },
          "id": {
            "example": "usr_2N9k8xQpL3vM5Rb1Tw7yZc",
            "type": "string"
          },
          "image": {
            "example": "https://lh3.googleusercontent.com/a/ACg8ocJ...",
            "nullable": true,
            "type": "string"
          },
          "name": {
            "example": "Tunde Adebayo",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "email"
        ],
        "type": "object"
      },
...
      "PublicInvitationDto": {
        "properties": {
          "email": {
            "example": "bola@example.com",
            "type": "string"
          },
          "event": {
            "$ref": "#/components/schemas/PublicInvitationEventDto"
          },
          "expiresAt": {
            "example": "2026-06-09T08:12:00.000Z",
            "type": "string"
          },
          "id": {
            "example": "b2c3d4e5-f6a7-4b8c-9d0e-1f2a3b4c5d6e",
            "format": "uuid",
            "type": "string"
          },
          "inviter": {
            "$ref": "#/components/schemas/PublicInvitationInviterDto"
          },
          "role": {
            "enum": [
              "admin",
              "contributor",
              "observer"
            ],
            "example": "contributor",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "accepted",
              "declined",
              "revoked",
              "expired"
            ],
            "example": "pending",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "status",
          "expiresAt",
          "event",
          "inviter"
        ],
        "type": "object"
      },
      "PublicInvitationEventDto": {
        "properties": {
          "eventType": {
            "example": "wedding",
            "type": "string"
          },
          "id": {
            "example": "6bb69700-72eb-40a0-b37e-e6c0b8b004a9",
            "format": "uuid",
            "type": "string"
          },
          "scheduledDate": {
            "description": "YYYY-MM-DD; null when no date is locked yet.",
            "example": "2026-06-24",
            "nullable": true,
            "type": "string"
          },
          "title": {
            "example": "Tunde & Bola — Wedding",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "eventType"
        ],
        "type": "object"
      },
      "PublicInvitationInviterDto": {
        "properties": {
          "email": {
            "example": "tunde@example.com",
            "type": "string"
          },
          "id": {
            "example": "usr_1A2b3C4d5E6f7G8h9I0jKl",
            "type": "string"
          },
          "name": {
            "example": "Tunde Adebayo",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "email"
        ],
        "type": "object"
      },
...
      "PublicRsvpInviteeDto": {
        "properties": {
          "customAnswers": {
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "dietary": "Vegetarian"
            },
            "nullable": true,
            "type": "object"
          },
          "email": {
            "example": "tunde@example.com",
            "type": "string"
          },
          "id": {
            "example": "6bb69700-72eb-40a0-b37e-e6c0b8b004a9",
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "example": "Tunde Olamide",
            "nullable": true,
            "type": "string"
          },
          "respondedAt": {
            "example": "2026-05-28T14:32:00.000Z",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "yes",
              "no",
              "maybe"
            ],
            "example": "pending",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "status"
        ],
        "type": "object"
      },
      "PublicRsvpResponseDto": {
        "properties": {
          "event": {
            "$ref": "#/components/schemas/PublicRsvpEventDto"
          },
          "invitee": {
            "$ref": "#/components/schemas/PublicRsvpInviteeDto"
          }
        },
        "required": [
          "invitee",
          "event"
        ],
        "type": "object"
      },
...
      "SubmitRsvpDto": {
        "properties": {
          "customAnswers": {
            "description": "Reserved for Phase 5c custom RSVP questions.",
            "type": "object"
          },
          "name": {
            "description": "Optional name correction — used when an invitee was added by email only.",
            "maxLength": 120,
            "type": "string"
          },
          "status": {
            "enum": [
              "yes",
              "no",
              "maybe"
            ],
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
...
      "UpdateMemberRoleDto": {
        "properties": {
          "role": {
            "enum": [
              "admin",
              "contributor",
              "observer"
            ],
            "type": "string"
          }
        },
        "required": [
          "role"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "cookie": {
        "in": "cookie",
        "name": "better-auth.session_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Event Planning Decision Making Assistant. Spec frozen per Phase 5.5 — bump `info.version` (major for breaking, minor for additive) on every committed change.",
    "title": "EPDEMA API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/events/{eventId}/members": {
      "get": {
        "operationId": "members_list",
        "parameters": [
          {
            "description": "Event UUID.",
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanningMemberDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List planning members of the event."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List planning members of an event",
        "tags": [
          "members"
        ]
      }
    },
    "/events/{eventId}/members/{memberId}": {
      "delete": {
        "operationId": "members_remove",
        "parameters": [
          {
            "description": "Event UUID.",
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Planning-member row UUID.",
            "in": "path",
            "name": "memberId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Remove a member from the event (admin only). No body."
          },
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a planning member (admin)",
        "tags": [
          "members"
        ]
      }
    },
    "/events/{eventId}/members/{memberId}/role": {
      "patch": {
        "operationId": "members_updateRole",
        "parameters": [
          {
            "description": "Event UUID.",
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Planning-member row UUID.",
            "in": "path",
            "name": "memberId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRoleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanningMemberDto"
                }
              }
            },
            "description": "Change a member's role (admin only)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change a planning member's role (admin)",
        "tags": [
          "members"
        ]
      }
    },
    "/events/{eventId}/members/invitations": {
      "get": {
        "operationId": "members_listInvitations",
        "parameters": [
          {
            "description": "Event UUID.",
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanningInvitationDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List pending invitations for the event (admin only)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List pending planning invitations (admin)",
        "tags": [
          "members"
        ]
      },
      "post": {
        "operationId": "members_invite",
        "parameters": [
          {
            "description": "Event UUID.",
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteMemberDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteMemberResponseDto"
                }
              }
            },
            "description": "Invite an email to the event (admin only). Response includes acceptUrl so admins can share manually when email delivery is unreliable."
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invite an email to the planning team (admin)",
        "tags": [
          "members"
        ]
      }
    },
    "/events/{eventId}/members/invitations/{invitationId}": {
      "delete": {
        "operationId": "members_revokeInvitation",
        "parameters": [
          {
            "description": "Event UUID.",
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Planning-invitation row UUID.",
            "in": "path",
            "name": "invitationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Revoke a pending invitation (admin only). No body."
          },
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke a pending planning invitation (admin)",
        "tags": [
          "members"
        ]
      }
    },
...
    "/invitations/{token}": {
      "get": {
        "operationId": "invitations_lookup",
        "parameters": [
          {
            "description": "Opaque invitation token (64-char hex) emailed to the recipient and returned in acceptUrl.",
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "example": "3f9c1e7b8a2d4f5e6c0b9a1d2e3f4a5b6c7d8e9f0a1b2c3d4e5f6a7b8c9d0e1f",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicInvitationDto"
                }
              }
            },
            "description": "Public peek at an invitation. Returns event title, inviter, expiry — no sensitive data."
          }
        },
        "summary": "Public peek at a planning invitation (unauthenticated)",
        "tags": [
          "invitations"
        ]
      }
    },
    "/invitations/{token}/accept": {
      "post": {
        "operationId": "invitations_accept",
        "parameters": [
          {
            "description": "Opaque invitation token (64-char hex) emailed to the recipient and returned in acceptUrl.",
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "example": "3f9c1e7b8a2d4f5e6c0b9a1d2e3f4a5b6c7d8e9f0a1b2c3d4e5f6a7b8c9d0e1f",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptInvitationResponseDto"
                }
              }
            },
            "description": "Accept an invitation. Requires signed-in user whose email matches."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accept a planning invitation (authenticated)",
        "tags": [
          "invitations"
        ]
      }
    },
...
    "/rsvp/{token}": {
      "get": {
        "operationId": "rsvp_lookup",
        "parameters": [
          {
            "description": "RSVP token from the invitation URL — base64url of 24 random bytes (~32 chars).",
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "example": "aBcD1234EFgh5678IJklMNop9012QrSt",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicRsvpResponseDto"
                }
              }
            },
            "description": "Look up an invitation by RSVP token. Returns event essentials, the invitee's own status, and (Production+ tier) the event brand so the RSVP page can render in the organizer's theme."
          }
        },
        "summary": "Look up an invitation by token",
        "tags": [
          "rsvp"
        ]
      },
      "post": {
        "operationId": "rsvp_submit",
        "parameters": [
          {
            "description": "RSVP token from the invitation URL — base64url of 24 random bytes (~32 chars).",
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "example": "aBcD1234EFgh5678IJklMNop9012QrSt",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitRsvpDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicRsvpResponseDto"
                }
              }
            },
            "description": "Submit / update an RSVP (yes / no / maybe). Idempotent — the same invitee can change their mind any time before the event."
          }
        },
        "summary": "Submit or update an RSVP",
        "tags": [
          "rsvp"
        ]
      }
    },
...
  },
  "servers": [],
...
}